+++ src/actions/mod.rs
@@ -1,0 +1,4 @@
+mod rescan;
+pub use rescan::rescan;
+
+pub mod elementary;